Don't have conflicting types with netdev_stdlib
The network_vlan and network_interface types are provided by the
netdev_stdlib module, so I've added that as a dependency in the
metadata.json. You don't want to have types named the same in both
modules.

Also, network_trunk and port_channel look like they are copied over but
have modifications. Any necessary modifications needed to work with
netdev_stdlib should be sent to netdev_stdlib as a pull request. I've
just renamed the files to preserve the changes.