Add baseline unit tests
These unit tests verify that each of the given manifests can compile.
They can be run with `bundle exec rake spec` . More documentation can
be found by searching the web for "rspec-puppet"

I couldn't get the tests to run; it appears that the ruby code requires
specific paths to files in /opt/puppetlabs/... and this inhibits any
portability. rubocop from the other commit should help sort the majority
of this out so that tests can run.